Create LICENSE
This license allows the user to commercially use and privately use, modify and distribute the codes. Moreover, the author has no liability on the discrepancies in the codes and does not provide any warranty.

Furthermore, the references must be cited if the codes will be used in any peer-reviewed publication.